Move add_prefix_to_rules function to constants.py for better organization
