+++ microSALT/utils/pubmlst/constants.py
@@ -1,6 +1,24 @@
 from enum import Enum
 from werkzeug.routing import Map, Rule
-from microSALT.utils.pubmlst.helpers import add_prefix_to_rules
+
+def add_prefix_to_rules(url_map: Map, prefix: str):
+    """
+    Add a prefix to all rules in the given URL map.
+
+    :param url_map: The original URL map.
+    :param prefix: The prefix to add (e.g., '/api').
+    :return: A new URL map with the prefix applied.
+    """
+    new_rules = []
+    new_rules.extend(
+        Rule(
+            f"{prefix}{rule.rule}",
+            endpoint=rule.endpoint,
+            methods=rule.methods,
+        )
+        for rule in url_map.iter_rules()
+    )
+    return Map(new_rules)
 
 CREDENTIALS_KEY: str = "credentials"
